audio subsystem: implement stubs + improve match percentages
Implement 20 audio stub functions and improve match percentages
via orchestrator-guided source fixes:

New 100% matches:
- Sound::SetPan (68.4→100%), RingBuffer::Read (72.2→100%),
  SampleInst::SynthPoll (0.2→100%), DelayEffect::Process (0.3→97.4%)

Major improvements:
- FlangerEffect::Process 0.2→77.5%, SampleData::Load 23.8→88.3%,
  SampleData::LoadWAV broken→54.3%, WavReader::Poll 0.7→82.3%,
  RingBuffer::Write 48.7→85.6%, MicNull::GetContinuousBuf 83→92.8%,
  EQEffect Reset/Process/SetParameter from stubs to 11-28%

Also includes ByteGrinder and VorbisReader improvements for native port.

+++ src/system/synth/ByteGrinder.h
@@ -5,11 +5,11 @@ class ByteGrinder {
 public:
     ByteGrinder() {}
     virtual ~ByteGrinder() {} // generic dtor
-    void GrindArray(long, long, unsigned char *, int, int);
+    void GrindArray(int, int, unsigned char *, int, int);
     void Init();
     void HvDecrypt(unsigned char *, unsigned char *, int);
 
 private:
-    unsigned long pickOneOf32A(bool, long);
-    unsigned long pickOneOf32B(bool, long);
+    unsigned int pickOneOf32A(bool, int);
+    unsigned int pickOneOf32B(bool, int);
 };
